refactor API deinition
